Simplification of form data configuration - subtypes are not used.
