+++ powerauth-nextstep-model/src/main/java/io/getlime/security/powerauth/lib/nextstep/model/entity/attribute/OperationFormFieldConfig.java
@@ -1,25 +1,18 @@
 package io.getlime.security.powerauth.lib.nextstep.model.entity.attribute;
 
-import com.fasterxml.jackson.annotation.JsonSubTypes;
-import com.fasterxml.jackson.annotation.JsonTypeInfo;
-
 import java.util.Objects;
 
 /**
- * Abstract class representing a form configuration field.
- *
- * This class requires JSON annotation since it is a base class that is extended with specific types.
+ * Operation form data configuration.
  *
  * @author Roman Strobl, roman.strobl@lime-company.eu
  *
  */
-@JsonTypeInfo(use = JsonTypeInfo.Id.NAME, include = JsonTypeInfo.As.PROPERTY, property = "type")
-@JsonSubTypes({
-        @JsonSubTypes.Type(value = OperationBankAccountChoiceFieldConfig.class, name = "BANK_ACCOUNT_CHOICE"),
-})
 public class OperationFormFieldConfig {
 
     protected String id;
+    private boolean enabled;
+    private String defaultValue;
 
     public String getId() {
         return id;
@@ -27,6 +20,22 @@ public class OperationFormFieldConfig {
 
     public void setId(String id) {
         this.id = id;
+    }
+
+    public boolean isEnabled() {
+        return enabled;
+    }
+
+    public void setEnabled(boolean enabled) {
+        this.enabled = enabled;
+    }
+
+    public String getDefaultValue() {
+        return defaultValue;
+    }
+
+    public void setDefaultValue(String defaultValue) {
+        this.defaultValue = defaultValue;
     }
 
     @Override
